requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
open3d>=0.13.0
scipy>=1.7.0
matplotlib>=3.4.0
PyYAML>=5.4.0
click>=8.0.0
tqdm>=4.62.0
scikit-image>=0.18.0
pillow>=8.3.0
setuptools>=61.0.0
wheel>=0.36.0
